+++ app/gui/src/project-view/util/codemirror/language/tableExpression/index.ts
@@ -34,6 +34,8 @@ export function useTableExpressionExtension(
           [
             ...suggestionDb.value.methods(EXPRESSION_STATICS_METHODS),
             ...suggestionDb.value.methods(COLUMN_METHODS),
+            ...suggestionDb.value.methods(NUMERIC_COLUMN_METHODS),
+            ...suggestionDb.value.methods(TEXT_COLUMN_METHODS),
           ].map((entry) => [entry.name, entry]),
         ).values(),
         methodInfoFromEntry,
@@ -58,17 +60,35 @@ const COLUMN_TYPE = ProjectPath.create(
   'Column.Column' as QualifiedName,
 )
 
+const NUMERIC_COLUMN_TYPE = ProjectPath.create(
+  'Standard.Table' as QualifiedName,
+  'Refined_Types.Numeric_Column.Numeric_Column' as QualifiedName,
+)
+
+const TEXT_COLUMN_TYPE = ProjectPath.create(
+  'Standard.Table' as QualifiedName,
+  'Refined_Types.Text_Column.Text_Column' as QualifiedName,
+)
+
 const EXPRESSION_STATICS_TYPE = ProjectPath.create(
   'Standard.Table' as QualifiedName,
   'Internal.Expression_Statics.Expression_Statics' as QualifiedName,
 )
 
-const EXPRESSION_STATICS_METHODS = {
-  memberOf: EXPRESSION_STATICS_TYPE,
-}
 const COLUMN_METHODS = {
   selfType: COLUMN_TYPE,
   name: (name: string) => !EXCLUDED_COLUMN_METHODS.has(name),
+}
+const NUMERIC_COLUMN_METHODS = {
+  selfType: NUMERIC_COLUMN_TYPE,
+  name: (name: string) => !EXCLUDED_COLUMN_METHODS.has(name),
+}
+const TEXT_COLUMN_METHODS = {
+  selfType: TEXT_COLUMN_TYPE,
+  name: (name: string) => !EXCLUDED_COLUMN_METHODS.has(name),
+}
+const EXPRESSION_STATICS_METHODS = {
+  memberOf: EXPRESSION_STATICS_TYPE,
 }
 
 function methodInfoFromEntry(entry: MethodSuggestionEntry): MethodCompletionInfo {
@@ -81,7 +101,6 @@ function methodInfoFromEntry(entry: MethodSuggestionEntry): MethodCompletionInfo
 }
 
 const EXCLUDED_COLUMN_METHODS = new Set([
-  'iif', // Used to implement `if`; no reason to call it as a method.
   'info', // Technically works, probably not useful.
   'rename', // When used with `Column.set`, this is redundant and doesn't work.
   'to_table', // Does nothing, successfully but inefficiently.